Don't use the asyncio loop's run_until_complete.
This makes the code harder to understand and modify.  Both the new and
the old code have a race condition: it's possible that the
reactor.listenTCP call will bind to the port and begin serving clients
before the MySQL connection is established.

Moving MyFactory.init out from MyFactory.__init__ will allow a future
refactoring to wait on the MySQL connection.